+++ src/Aggregator/Cms/CmsBlockCountAggregator.php
@@ -14,9 +14,9 @@ use RunAsRoot\PrometheusExporter\Api\MetricAggregatorInterface;
 use RunAsRoot\PrometheusExporter\Service\UpdateMetricService;
 use Magento\Cms\Api\BlockRepositoryInterface;
 
-class CmsBlocksCountAggregator implements MetricAggregatorInterface
+class CmsBlockCountAggregator implements MetricAggregatorInterface
 {
-    private const METRIC_CODE = 'magento2_cms_block_count_total';
+    private const METRIC_CODE = 'magento_cms_block_count_total';
 
     /**
      * @var UpdateMetricService
@@ -65,10 +65,8 @@ class CmsBlocksCountAggregator implements MetricAggregatorInterface
     public function aggregate(): bool
     {
         $searchCriteria = $this->searchCriteriaBuilder->create();
-
         $cmsSearchResult = $this->cmsRepository->getList($searchCriteria);
-        $this->updateMetricService->update(self::METRIC_CODE, (string)$cmsSearchResult->getTotalCount());
 
-        return true;
+        return $this->updateMetricService->update(self::METRIC_CODE, (string)$cmsSearchResult->getTotalCount());
     }
 }