+++ src/Aggregator/Customer/CustomerCountAggregator.php
@@ -1,5 +1,4 @@
 <?php
-
 declare(strict_types=1);
 /**
  * @copyright see PROJECT_LICENSE.txt
@@ -7,29 +6,24 @@ declare(strict_types=1);
  * @see PROJECT_LICENSE.txt
  */
 
-namespace RunAsRoot\PrometheusExporter\Aggregator\Order;
+namespace RunAsRoot\PrometheusExporter\Aggregator\Customer;
 
+use Magento\Customer\Api\CustomerRepositoryInterface;
 use Magento\Framework\Api\SearchCriteriaBuilder;
-use Magento\Framework\Exception\CouldNotSaveException;
+use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NoSuchEntityException;
-use Magento\Sales\Api\OrderRepositoryInterface;
 use Magento\Store\Api\StoreRepositoryInterface;
 use RunAsRoot\PrometheusExporter\Api\MetricAggregatorInterface;
 use RunAsRoot\PrometheusExporter\Service\UpdateMetricService;
 
-class OrderCountAggregator implements MetricAggregatorInterface
+class CustomerCountAggregator implements MetricAggregatorInterface
 {
-    private const METRIC_CODE = 'magento2_orders_count_total';
+    private const METRIC_CODE = 'magento2_customer_count_total';
 
     /**
      * @var UpdateMetricService
      */
     private $updateMetricService;
-
-    /**
-     * @var OrderRepositoryInterface
-     */
-    private $orderRepository;
 
     /**
      * @var SearchCriteriaBuilder
@@ -41,14 +35,19 @@ class OrderCountAggregator implements MetricAggregatorInterface
      */
     private $storeRepository;
 
+    /**
+     * @var CustomerRepositoryInterface
+     */
+    private $customerRepository;
+
     public function __construct(
         UpdateMetricService $updateMetricService,
-        OrderRepositoryInterface $orderRepository,
+        CustomerRepositoryInterface $customerRepository,
         StoreRepositoryInterface $storeRepository,
         SearchCriteriaBuilder $searchCriteriaBuilder
     ) {
         $this->updateMetricService = $updateMetricService;
-        $this->orderRepository = $orderRepository;
+        $this->customerRepository = $customerRepository;
         $this->searchCriteriaBuilder = $searchCriteriaBuilder;
         $this->storeRepository = $storeRepository;
     }
@@ -60,7 +59,7 @@ class OrderCountAggregator implements MetricAggregatorInterface
 
     public function getHelp(): string
     {
-        return 'Magento2 Order Count by state';
+        return 'Magento2 Customer Count by state';
     }
 
     public function getType(): string
@@ -69,26 +68,23 @@ class OrderCountAggregator implements MetricAggregatorInterface
     }
 
     /**
-     * @return bool
-     * @throws CouldNotSaveException
-     *
+     * @throws LocalizedException
      */
     public function aggregate(): bool
     {
         $searchCriteria = $this->searchCriteriaBuilder->create();
 
-        $orderSearchResult = $this->orderRepository->getList($searchCriteria);
+        $searchResult = $this->customerRepository->getList($searchCriteria);
 
-        if ($orderSearchResult->getTotalCount() === 0) {
+        if ($searchResult->getTotalCount() === 0) {
             return true;
         }
 
-        $orders = $orderSearchResult->getItems();
+        $customers = $searchResult->getItems();
 
         $countByStore = [];
-        foreach ($orders as $order) {
-            $state = $order->getState();
-            $storeId = $order->getStoreId();
+        foreach ($customers as $customer) {
+            $storeId = $customer->getStoreId();
 
             try {
                 $store = $this->storeRepository->getById($storeId);
@@ -98,22 +94,16 @@ class OrderCountAggregator implements MetricAggregatorInterface
             }
 
             if (!array_key_exists($storeCode, $countByStore)) {
-                $countByStore[$storeCode] = [];
+                $countByStore[$storeCode] = 0;
             }
 
-            if (!array_key_exists($state, $countByStore)) {
-                $countByStore[$storeCode][$state] = 0;
-            }
-
-            $countByStore[$storeCode][$state]++;
+            $countByStore[$storeCode]++;
         }
 
-        foreach ($countByStore as $storeCode => $countByState) {
-            foreach ($countByState as $state => $count) {
-                $labels = ['state' => $state, 'store_code' => $storeCode];
+        foreach ($countByStore as $storeCode => $count) {
+            $labels = ['store_code' => $storeCode];
 
-                $this->updateMetricService->update(self::METRIC_CODE, (string)$count, $labels);
-            }
+            $this->updateMetricService->update(self::METRIC_CODE, (string)$count, $labels);
         }
 
         return true;