+++ src/Test/Unit/Aggregator/Cms/CmsBlockCountAggregatorUnitTest.php
@@ -8,16 +8,18 @@
 namespace RunAsRoot\PrometheusExporter\Test\Unit\Aggregator\Cms;
 
 use Magento\Cms\Api\BlockRepositoryInterface;
+use Magento\Cms\Api\Data\BlockSearchResultsInterface;
+use Magento\Framework\Api\SearchCriteria;
 use Magento\Framework\Api\SearchCriteriaBuilder;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use RunAsRoot\PrometheusExporter\Aggregator\Cms\CmsBlocksCountAggregator;
+use RunAsRoot\PrometheusExporter\Aggregator\Cms\CmsBlockCountAggregator;
 use RunAsRoot\PrometheusExporter\Service\UpdateMetricService;
 
 class CmsBlockCountAggregatorUnitTest extends TestCase
 {
     /**
-     * @var CmsBlocksCountAggregator
+     * @var CmsBlockCountAggregator
      */
     private $sut;
 
@@ -25,23 +27,31 @@ class CmsBlockCountAggregatorUnitTest extends TestCase
     {
         parent::setUp();
 
-        /** @var UpdateMetricService | MockObject $updateMetricService */
-        $updateMetricService = $this->createMock(UpdateMetricService::class);
-        /** @var BlockRepositoryInterface | MockObject $cmsRepository */
-        $cmsRepository = $this->createMock(BlockRepositoryInterface::class);
-        /** @var SearchCriteriaBuilder | MockObject $searchCriteriaBuilder */
-        $searchCriteriaBuilder = $this->createMock(SearchCriteriaBuilder::class);
+        /** @var SearchCriteria | MockObject $searchCriteria */
+        $searchCriteria = $this->getMockBuilder(SearchCriteria::class)->disableOriginalConstructor()->getMock();
 
-        $this->sut = new CmsBlocksCountAggregator(
+        /** @var UpdateMetricService | MockObject $updateMetricService */
+        $updateMetricService = $this->getMockBuilder(UpdateMetricService::class)->disableOriginalConstructor()->getMock();
+        $updateMetricService->method('update')->willReturn(true);
+
+        $searchResultInterface = $this->getMockBuilder(BlockSearchResultsInterface::class)->getMockForAbstractClass();
+        $searchResultInterface->expects($this->once())->method('getTotalCount')->willReturn('10');
+
+        /** @var BlockRepositoryInterface | MockObject $cmsRepository */
+        $cmsRepository = $this->getMockBuilder(BlockRepositoryInterface::class)->getMockForAbstractClass();
+        $cmsRepository->method('getList')->willReturn($searchResultInterface);
+
+        /** @var SearchCriteriaBuilder | MockObject $searchCriteriaBuilder */
+        $searchCriteriaBuilder = $this->getMockBuilder(SearchCriteriaBuilder::class)->disableOriginalConstructor()->getMock();
+        $searchCriteriaBuilder->method('create')->willReturn($searchCriteria);
+
+        $this->sut = new CmsBlockCountAggregator(
             $updateMetricService,
             $cmsRepository,
             $searchCriteriaBuilder
         );
     }
 
-    /**
-     * @throws \Magento\Framework\Exception\LocalizedException
-     */
     public function testExecuteReturnPrometheusResult(): void
     {
         $this->assertTrue($this->sut->aggregate());